fix darknet2ncnn.cpp warnings (#2209)
* fix darknet2ncnn.cpp warnings

* revoke extra warning check

+++ tools/darknet/darknet2ncnn.cpp
@@ -828,7 +828,7 @@ void load_weights(const char* filename, std::deque<Section*>& dnet)
             }
 
             if (s->layer_type == "Convolution")
-                read_to(s->weights, s->c * s->filters * s->size * s->size, fp);
+                read_to(s->weights, (size_t)(s->c) * s->filters * s->size * s->size, fp);
             else if (s->layer_type == "ConvolutionDepthWise")
                 read_to(s->weights, s->c * s->filters * s->size * s->size / s->groups, fp);
         }
